backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==1.10.13
python-multipart==0.0.6
asyncio-mqtt==0.16.2
python-dotenv>=1.0.0,<1.1.0
typing-extensions==4.9.0
starlette==0.27.0
anyio==3.7.1
click==8.1.7
h11==0.14.0
httptools==0.6.1
python-json-logger==2.0.7
watchfiles==0.21.0
paho-mqtt==1.6.1
psutil==5.9.5
random2==1.0.1
time-machine==2.10.0